[direct] move shiftedWhile test to KyoTest
